% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spader.R
\name{ChaoSpecies}
\alias{ChaoSpecies}
\title{Estimation of species richness}
\usage{
ChaoSpecies(data, datatype = c("abundance", "incidence"), k = 10,
  conf = 0.95)
}
\arguments{
\item{data}{a column vector of species abundances or incidence frequencies.\cr
If \code{datatype = "incidence"}, then the first entry of the input data must be total number of sampling units,
followed by species incidence frequencies. (See example \code{data(ChaoSpeciesDemoInci)}).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or
sampling-unit-based incidence data (\code{datatype = "incidence"}).}

\item{k}{cut-off point; it is a value that separates frequency counts into abundant and rare groups.}

\item{conf}{a positive number \eqn{\le} 1 specifying the level of confidence interval.}
}
\value{
a list of three objects:
\code{$Basic.Data.Information} and \code{$Rare.Species.Group}/\code{$Infreq.Species.Group}
for summarizing data information; \code{$Species.Table} for showing a table of various species richness estimates, standard errors,
and the associated confidence intervals.
}
\description{
\code{ChaoSpecies}: estimation of species richness in one assemblage
}
\examples{
data(ChaoSpeciesDemoAbu)
ChaoSpecies(ChaoSpeciesDemoAbu, datatype="abundance", k = 10, conf = 0.95)
data(ChaoSpeciesDemoInci)
ChaoSpecies(ChaoSpeciesDemoInci, datatype="incidence", k = 10, conf = 0.95)
}
\references{
Burnham, K. P. and Overton, W. S. (1978). Estimation of the size of a closed population when capture probabilities vary among animals. Biometrika 65, 625-633.\cr\cr
Chao, A. (1984). Nonparametric estimation of the number of classes in a population. Scandinavian Journal of Statistics 11, 265-270.\cr\cr
Chao, A. (1987). Estimating the population size for capture-recapture data with unequal catchability. Biometrics 43, 783-791.\cr\cr
Chao, A. (2005). Species estimation and applications. Encyclopedia of Statistical Sciences, Second Edition, Vol. 12, 7907-7916 (N. Balakrishnan, C. B. Read and B. Vidakovic, Editors), Wiley, New York.\cr\cr
Chao, A., Hwang, W.-H., Chen, Y.-C. and Kuo. C.-Y. (2000). Estimating the number of shared species in two communities. Statistica Sinica 10, 227-246.\cr\cr
Chao, A. and Lee, S.-M. (1992). Estimating the number of classes via sample coverage. Journal of the American Statistical Association 87, 210-217.\cr\cr
Chao, A., Ma, M-C. and Yang, M. C. K. (1993). Stopping rule and estimation for recapture debugging with unequal detection rates. Biometrika 80, 193-201.\cr\cr
Chao, A., Shen, T.-J. and Hwang, W. H. (2006). Application of Laplace boundary-mode approximations to estimate species and shared species richness. Australian and New Zealand Journal of Statistics 48, 117-128.\cr\cr
Chiu, C.-H., Wang Y. T., Walther B. A. and Chao A. (2014). An improved non-parametric lower bound of species richness via the Good-Turing frequency formulas. Biometrics 70, 671-682.\cr\cr
Lee, S.-M. and Chao, A. (1994). Estimating population size via sample coverage for closed capture-recapture models. Biometrics 50, 88-97.\cr\cr
Shen, T.-J. (2003). Prediction of Biodiversity, Ph.D. dissertation, National Tsing-Hua University, HsinChu, Taiwan.\cr\cr
Shen, T.-J., Chao, A. and Lin, J.-F. (2003). Predicting the number of new species in further taxonomic sampling. Ecology 84, 798-804.
}

