\name{DiversityDataAbu}
\alias{DiversityDataAbu}
\docType{data}
\title{
 Rain Forest Data for Diversity
}
\description{
 Old-growth rain forest data was collected by Dr. Robin Chazdon and colleagues, and  discussed in Chao et al. (2008). It contains total number of individuals 557 and 69 observed species. The first species has abundance 17, the second species has abundance 14, and so on.
}
\usage{data(DiversityDataAbu)}
\format{
 The data input format (a column vector with 120 entries) is:\cr\cr
    17 \cr
    14 \cr
   121 \cr
     . \cr  
     . \cr
     . \cr
     1 \cr
     1 \cr
     1
}

\references{
  Chao, A., Jost, L., Chiang, S. C., Jiang, Y.-H. and Chazdon, R. (2008). A Two
  -stage probabilistic approach to multiple-community similarity indices. 
  Biometrics 64, 1178-1186.
}

\keyword{datasets}
