\name{DiversityDataAbu_count}
\alias{DiversityDataAbu_count}
\docType{data}
\title{
Beetles Data for Diversity
}
\description{
Tropical foliage insects from sweep samples in Costa Rica. Here the most abundant beetle species were represented by 42 individuals. Here the first pair (1, 84) indicates that there are 84 singletons, the second pair (2, 10) indicates there are 10 doubletons, and so on.
}
\usage{data("DiversityDataAbu_count")}
\format{
  The data input format is:\cr\cr
  1 \cr
  84\cr
  2 \cr
  10\cr
  . \cr
  . \cr
  . \cr
  . \cr
  .\cr
  42\cr
  1
}
\references{
Janzen, D.H. (1973) Sweep samples of tropical foliage insects: description of study sites, with data on species abundances and size distributions. Ecology, 54, 659-686.
}
\keyword{datasets}
