\name{DiversityDataInci}
\alias{DiversityDataInci}
\docType{data}
\title{
Ant Data for Diversity
}
\description{
The data includes incidence frequencies of tropical rainforest ants using sampling techniques: Berlese extraction of soil samples (217 samples). It includes 117 observed species.
}
\usage{data("DiversityDataInci")}
\format{
  The data input format (a column vector with 277 entries) is:\cr\cr
    217 \cr
    1   \cr
    0   \cr
    .   \cr
    .   \cr
    .   \cr
    0   \cr
    1   \cr
    0
}
\references{
Longino, J.T., Coddington, J. and Colwell, R.K. (2002). The ant fauna of a tropical rain forest: estimating species richness three different ways. Ecology, 83, 689-702.
}
\keyword{datasets}
