\name{DiversityDataInci_freq_count}
\alias{DiversityDataInci_freq_count}
\docType{data}
\title{
Seed-bank Data for Diversity
}
\description{
The soil seed bank was studied in a gallery forest in Southeastern Brazil. The first entry, indicating that there are 121 sampling units, is followed by (1, 3), (2, 2),..., and so on. Here (1, 3) indicates that there are 3 unique species, (2, 2) indicates there are 2 duplicate species, and so on.
}
\usage{data("DiversityDataInci_freq_count")}
\format{
  The data input format is:\cr\cr
  121 \cr
  1 \cr
  1 \cr
  1 \cr
  2 \cr
  2 \cr
  . \cr
  . \cr
  61 \cr
  1
}
\references{
Colwell, R. K. and Coddington, J. A. (1994). Estimating terrestrial biodiversity through extrapolation. Philosophical Transactions of the Royal Society of London B - Biological Sciences 345, 101-118. 
}
\keyword{datasets}
