\name{DiversityDataInci_raw}
\alias{DiversityDataInci_raw}
\docType{data}
\title{
Seed-bank Data for Diversity
}
\description{
The soil seed bank was studied in a gallery forest in Southeastern Brazil. There are 121 soil sampling units and 34 observed species.
}
\usage{data("DiversityDataInci_raw")}
\format{
  The data input format (a data.frame with 34 rows and 121 columns) is: \cr\cr
  1 1 1 1 0 . . . \cr
  1 0 0 0 0 . . . \cr
  1 1 1 0 0 . . . \cr
  1 0 1 0 0 . . . \cr
  1 0 0 0 1 . . . \cr
  . . . . . \cr
  . . . . . \cr
  . . . . . 
}
\references{
Colwell, R. K. and Coddington, J. A. (1994). Estimating terrestrial biodiversity through extrapolation. Philosophical Transactions of the Royal Society of London B - Biological Sciences 345, 101-118. 
}
\keyword{datasets}
