% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spader.R
\name{Diversity}
\alias{Diversity}
\title{Estimation of species diversity}
\usage{
Diversity(data, datatype = c("abundance", "incidence"), q = NULL)
}
\arguments{
\item{data}{a column vector of species abundances or incidence frequencies.\cr
If \code{datatype = "incidence"}, then the first entry of the input data must be total number of sampling units,
followed by species incidence frequencies. (See example \code{data(ChaoSpeciesDataInci)}).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or
sampling-unit-based incidence data (\code{datatype = "incidence"}).}

\item{q}{a numeric value specifying the diversity order of Hill numbers.}
}
\value{
a list of seven objects:
\code{$BASIC.DATA} for summarizing data information; \code{$SPECIES.RICHNESS} for showing various species richness estimates along with related statistics;
\code{$SHANNON.INDEX} and \code{$EXPONENTIAL.OF.SHANNON.INDEX} for showing various Shannon index/diversity estimates;
\code{$SIMPSON.INDEX} and \code{$INVERSE.OF.SIMPSON.INDEX} for showing various Simpson index/diversity estimates;
\code{$HILL.NUMBERS} for showing Hill number (diversity of order from 0 to 3) estimates. \cr\cr
}
\description{
\code{Diversity}: estimation of various diversity indices including species richness (diversity of order 0),
the Shannon index/diversity (diversity of order 1), the Simpson index/diversity (diversity of order 2),
and Hill number (diversity of order from 0 to 3) .
}
\examples{
\dontrun{
data(DiversityDataAbu)
Diversity(DiversityDataAbu,datatype="abundance")
data(ChaoSpeciesDataInci)
Diversity(ChaoSpeciesDataInci,datatype="incidence")
}
}
\references{
Chao, A. (1984). Nonparametric estimation of the number of classes in a population. Scandinavian Journal of Statistics 11, 265-270.\cr\cr
Chao, A. (1987). Estimating the population size for capture-recapture data with  unequal catchability. Biometrics 43, 783-791.\cr\cr
Chao, A. (2005). Species estimation and applications. Encyclopedia of Statistical Sciences, Second Edition, Vol. 12, 7907-7916 (N. Balakrishnan, C. B. Read and B. Vidakovic, Editors), Wiley, New York.\cr\cr
Chao, A., Hwang, W.-H., Chen, Y.-C. and Kuo. C.-Y. (2000). Estimating the number of shared species in two communities. Statistica Sinica 10, 227-246.\cr\cr
Chao, A. and Jost, L. (2015). Estimating diversity and entropy profiles via discovery rates of new species. Methods in Ecology and Evolution.\cr\cr
Chao, A. and Lee, S.-M. (1992). Estimating the number of classes via sample coverage. Journal of the American Statistical Association 87, 210-217.\cr\cr
Chao, A., Ma, M-C. and Yang, M. C. K. (1993). Stopping rule and estimation for recapture debugging with unequal detection rates. Biometrika 80, 193-201.\cr\cr
Chao, A. and Shen, T.-J. (2003). Nonparametric estimation of Shannon index of diversity when there are unseen species in sample. Environmental and Ecological Statistics 10, 429-443.\cr\cr
Chao, A., Shen, T.-J. and Hwang, W. H. (2006). Application of Laplace boundary-mode approximations to estimate species and shared species richness. Australian and New Zealand Journal of Statistics 48, 117-128.\cr\cr
Chao, A., Wang, Y. T. and Jost, L. (2013). Entropy and the species accumulation curve: a novel estimator of entropy via discovery rates of new species. Methods in Ecology and Evolution 4, 1091-1110.\cr\cr
Magurran, A. E. (1988). Ecological Diversity and Its Measurement. Princeton, Princeton University Press, New Jersey.\cr\cr
Shen, T.-J., Chao, A. and Lin, J.-F. (2003). Predicting the number of new species in further taxonomic sampling. Ecology 84, 798-804.\cr\cr
Zahl, S. (1977). Jackknifing an index of diversity. Ecology 58, 907-913.
}

