\name{GeneticDataAbu}
\alias{GeneticDataAbu}
\docType{data}
\title{
Human Allele Frequency Data for Genetic
}
\description{
The data is allele frequency counts from four human subpopulations (BiakaPyg, Palestin, Bedouin and Druze) taken from the data set provided by Rosenberg et al. (2002). Beginning with the first row,
each row denotes the  number of frequency count of each allele  in all the samples. In this demo
data, the first allele was observed 0, 6, 1, 1, respectively. The second allele was observed 11, 0, 0, 0,
respectively, and so on.
}
\usage{data("GeneticDataAbu")}
\format{
 The data input format (a matrix with 27 rows and 4 columns) is:\cr\cr
  0 6 1 1 \cr
 11 0 0 0 \cr
  5 0 0 0 \cr
  . . . . \cr
  . . . . \cr
  . . . . \cr
  0 0 0 0 
}


\references{
Rosenberg, N. A., Pritchard, J. K., Weber, J. L., Cann, H. M., Kidd, K. K., Zhivotovsky, L. A. and Feldman, M. W. (2002). Genetic structure of human populations. Science 298, 2381-2385.
}

\keyword{datasets}
