\name{GeneticsDataAbu}
\alias{GeneticsDataAbu}
\docType{data}
\title{
Human allele frequency data for  Function Genetics
}
\description{
The data taken from Rosenberg et al. (2002) consist of allele frequencies from four human
subpopulations (BiakaPyg, Palestin, Bedouin and Druze). The data are formatted as an allele
(row) by subpopulation (column) matrix file. Entries in each row denote the frequencies of each
allele found in the four subpopulations. The data include an observed allele frequency table with 27 rows and 4 columns.
}
\usage{data(GeneticsDataAbu)}

\references{
Rosenberg, N. A., Pritchard, J. K., Weber, J. L., Cann, H. M., Kidd, K. K., Zhivotovsky, L. A. and Feldman, M. W. (2002). Genetic structure of human populations. Science, 298, 2381-2385.
}
\keyword{datasets}
