\name{ChaoSharedDataAbu}
\alias{ChaoSharedDataAbu}
\docType{data}
\title{
 Birds data for ChaoShared
}
\description{
  The data is bird abundances/frequencies records of two estuaries (Chao et al. 2000). Total of 85867 and 59646 individuals were separately observed from two estuaries. In these two estuaries, there were 155 and 140 species observed respectively, where 111 species are shared. For each species (row), the entry of the first column is the observed species frequency (or abundance) from estuary I, and the second column is the observed species frequency (or abundance) from estuary II.
}
\usage{data(ChaoSharedDataAbu)}
\format{
 The data input format (a data.frame with 201 rows and 2 columns) is:\cr\cr
 39 42 \cr
  0 70 \cr
  0  1 \cr
  .  . \cr
  .  . \cr
  .  . \cr
  1  1 
}
\references{
 Chao, A., Hwang, W.-H., Chen, Y.-C. and Kuo. C.-Y. (2000). Estimating the number of shared species in two communities. Statistica Sinica 10, 227-246.
}

\keyword{datasets}
