\name{ChaoSharedDemoInci}
\alias{ChaoSharedDemoInci}
\docType{data}
\title{
 Birds data for ChaoShared
}
\description{
 The data is bird incidence frequencies of two assemblages (Chao, Shen and Hwang 2006). The first row denotes the number of sampling units 19 and 20, respectively. Beginning with the second row, each row denotes the total number of presences of each species in all the samples. In this data, the first and second species in the bird list were not observed, and the species in the bottom row was observed by 19 and 20, respectively. 
 }
\usage{data(ChaoSharedDemoInci)}
\format{
 The data input format (a matrix/data.frame with 359 rows and 2 columns) is:\cr\cr
 19 20\cr
  0 0 \cr
  0 0 \cr
  . . \cr  
  . . \cr
  . . \cr
 19 20 
}
\references{
 Chao, A., Shen, T.-J. and Hwang, W. H. (2006). Application of Laplace boundary-mode approximations to estimate species and shared species richness. Australian and New Zealand Journal of Statistics 48, 117-128.
}
\keyword{datasets}
