\name{ChaoSharedDataInci_raw}
\alias{ChaoSharedDataInci_raw}
\docType{data}
\title{
The Hong Kong Big Bird Race Data for ChaoShared
}
\description{
In 2015, a total of 223 species was observed by 16 teams competed in Hong
Kong Big Bird Race. In 2016, same number of species was observed by 17 teams. Merging the two-year data by species names, there were 280 species in the birds checklist.
}
\usage{data("ChaoSharedDataInci_raw")}
\format{
A list 2: \cr
For the first list (\code{ChaoSharedDataInci_raw[[1]]}), a data.frame with 280 rows and 16 columns  ; \cr 
For the second list (\code{ChaoSharedDataInci_raw[[2]]}), a data.frame with 280 rows and 17 columns.\cr\cr
0 0 0 0 0 . . . \cr
0 0 0 0 0 . . . \cr
0 0 0 0 0 . . . \cr
1 0 0 0 0 . . . \cr
0 0 0 1 0 . . . \cr
. . . . . \cr
. . . . . \cr
. . . . . \cr\cr

0 0 0 0 0 . . . \cr
0 0 0 0 0 . . . \cr
0 0 0 0 0 . . . \cr
0 0 0 0 1 . . . \cr
0 0 0 1 1 . . . \cr
. . . . . \cr
. . . . . \cr
. . . . . 

}
\references{
WWF-Hong Kong's longest-running fundraising event. Established in 1984.
}
\keyword{datasets}
