\name{SimilarityMultDataInci}
\alias{SimilarityMultDataInci}
\docType{data}
\title{
Ant Data for SimilarityMult
}
\description{
The data was collected in Costa-Rica (Longino et al. 2002). It includes incidence frequencies of tropical rainforest ants using three sampling techniques: (a) Berlese extraction of soil samples (217 samples), (b) fogging samples from canopy fogging (459 samples), and (c) Malaise trap samples for flying and crawling insects (62 samples). the first row denotes the number of sampling units (217, 459 and 62 denote the number of sampling units for Berlese, fogging, and Malaise, respectively). Beginning with the second row, the three numbers in each row denotes respectively the total number of detections of a given species in 217 Berlese samples, 459 fogging samples, and 62 Malaise samples.
}
\usage{data("SimilarityMultDataInci")}
\format{
The data input format (a matrix with 277 rows and 3 columns) is:\cr\cr
  217 459 62  \cr
   1    0  0  \cr
   0   27  3  \cr
   .    .  .  \cr  
   .    .  .  \cr
   .    .  .  \cr
   0   12  0 
}
\references{
Longino, J. T., Coddington, J. and Colwell, R. K. (2002). The ant fauna of a tropical rain forest: estimating species richness three different ways. Ecology 83, 689-702.
}
\keyword{datasets}
