\name{SimilarityMultDataInci_raw}
\alias{SimilarityMultDataInci_raw}
\docType{data}
\title{
Soil Ciliates Data for SimilarityMult
}
\description{
A total of 51 soil samples were taken from three areas ( EtoshaPan, CentralNamibDesert, SouthernNamibDesert) of Namibia. In area EtoshaPan, there were 19 soil samples and the number of observed species was 216. In area CentralNamibDesert, there were 17 soil samples and the number of observed species was 130. In area SouthernNamibDesert, there were 15 soil samples and the number of observed species was 150. Merging the three data by species names, there were 365 species in the Ciliates checklist.
}
\usage{data("SimilarityMultDataInci_raw")}
\format{
A list 3: \cr\cr
For the first list (\code{SimilarityMultDataInci_raw[[1]]}), a matrix with 365 rows and 19 columns; \cr
For the second list (\code{SimilarityMultDataInci_raw[[2]]}), a matrix with 365 rows and 17 columns;\cr
For the third list (\code{SimilarityMultDataInci_raw[[3]]}), a matrix with 365 rows and 15 columns. \cr\cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr\cr

  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr\cr
  
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 1 0 0 0 . . . \cr
  0 0 0 0 0 . . . 
}

\references{
Foissner, W., Agatha, S., & Berger, H. (2002) Soil Ciliates (Protozoa, Ciliophora) From Namibia (Southwest Africa), With Emphasis on Two Contrasting Environments, the Etosha Region and the Namib Desert. Denisia, 5, 1-1459.
}
\keyword{datasets}
