\name{SpadeR-package}
\alias{SpadeR-package}
\alias{SpadeR}
\docType{package}
\title{Species prediction and diversity estimation with R

}
\description{
Computing various biodiversity indices based on individual-based (abundance) data or sampling-unit-based (incidence) data taken from one or multiple assemblages. This package contains six parts: \cr\cr
1.ChaoSpecies (estimating species richness for one assemblage). \cr\cr
2.Diversity (estimating a continuous diversity profile including species richness, Shannon index/diversity and Simpson index/diversity). \cr\cr
3.ChaoShared (estimating the number of shared species for two assemblages). \cr\cr
4.SimilartyPair (estimating various similarity indices for two assemblages). The sampling-unit-based indices include the class Jaccard and Sorensen indices; the individual-based indices include the Bray-curtis, Morisita-Horn, Horn and abundance-based Jaccard and Sorensen indices. \cr\cr 
5.SimilarityMult (estimating Sorensen, Horn, and Morisita similarity indices for more than two assemblages). \cr\cr
6.Genetic (estimating allelic differentiation/similarity among subpopulations). This part is almost identical to SimilarityMult except that it is focused on genetic applications.
}
\details{
\tabular{ll}{
Package: \tab SpadeR\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2016-07-20\cr
License: \tab GPL (>= 3)\cr
}
functions: ChaoSpecies, Diversity, ChaoShared, SimilarityPair,
SimilarityMult,  Genetic
}
\author{
Anne Chao, K. H. Ma and T. C. Hsieh 

Maintainer: Anne Chao <chao@stat.nthu.edu.tw>

}
\keyword{ package }


