\name{ChaoSpeciesDataAbu}
\alias{ChaoSpeciesDataAbu}
\docType{data}
\title{
 Bird species abundances/frequencies for ChaoSpecies
}
\description{
 The data is birds abundances/frequencies from Magurran (1988, p.152). It contains 25 observed species. The first species has abundance 752, the second species has abundance 276, and so on.
}
\usage{data(ChaoSpeciesDataAbu)}
\format{
  The data input format (a column vector with 25 observations) is:\cr\cr
  752 \cr
  276 \cr
  194 \cr
   .  \cr
   .  \cr
   .  \cr
   1  \cr
   1  \cr
   1
}

\references{
  Magurran, A. E. (1988). Ecological Diversity and Its Measurement. 
  Princeton, Princeton University Press, New Jersey. 
}

\keyword{datasets}
