\name{ChaoSpeciesDataAbu_count}
\alias{ChaoSpeciesDataAbu_count}
\docType{data}
\title{
Vascular Plant Data for ChaoSpecies
}
\description{
The data is endangered and rare vascular plant species in the central portion of the southern Appalachian region. A total of 188 species were recorded out of 1008 individuals compiled over a span of 150 years of field observations. Here the first pair (1, 61) indicates that there are 61 singletons, the second pair (2, 35) indicates there are 35 doubletons, and so on.
}
\usage{data("ChaoSpeciesDataAbu_count")}
\format{
  The data input format is:\cr\cr
  1 \cr
  61\cr
  2 \cr
  35\cr
  3 \cr
  18\cr
  4 \cr
  12  \cr
  5 \cr
  15\cr
  .\cr
  .\cr
  .
}
\references{
Miller, R. I., and R. G. Wiegert. 1989. Documenting completeness, species-area relations, and the species-abundance distribution of a regional flora. Ecology 70,16-22
}
\keyword{datasets}
