\name{ChaoSpeciesDataInci}
\alias{ChaoSpeciesDataInci}
\docType{data}
\title{
  Seed-bank Data for ChaoSpecies
}
\description{
  The data is a set of soil samples of seedlings data (Colwell and Coddington 1994). It contains 34 observed species. The  first entry denotes the number of sampling units 121. Beginning with the second row, each row denotes the total number of presences of each species in the sample. In this data, the first species in 121 sampling units was observed 61, the second species in 121 sampling units was observed 47, and so on. 
}
\usage{data(ChaoSpeciesDataInci)}
\format{
 The data input format (a column vector with 35 entries)is:\cr\cr
  121 \cr
   61 \cr
   47 \cr
   52 \cr  
   .  \cr
   .  \cr
   .  \cr
   2  \cr
   1  \cr
   1

}
\references{
  Colwell, R. K. and Coddington, J. A. (1994). Estimating terrestrial biodiversity through extrapolation. Philosophical Transactions of the Royal Society of London B-Biological Sciences 345, 101-118. 
}

\keyword{datasets}
