\name{ChaoSpeciesDataInci_freq_count}
\alias{ChaoSpeciesDataInci_freq_count}
\docType{data}
\title{
Cotton Tail Rabbit Data for ChaoSpecies
}
\description{
Live trapping was conducted for 18 consecutive nights. Marking or tagging is needed to distinguish individuals. Often, a unique tag was attached to each individual so that the capture history of each individual captured in the experiment was known. The first entry, indicating that there are 18 samples, is followed by (1, 43), (2, 16),..., and so on. Here (1, 43) indicates that there are 43 unique species, (2, 16) indicates there are 16 duplicate species, and so on.
}
\usage{data("ChaoSpeciesDataInci_freq_count")}
\format{
  The data input format is:\cr\cr
  18  \cr
  1 \cr
  43  \cr
  2 \cr
  16  \cr
  3 \cr
  8 \cr
  4 \cr
  6 \cr
  5 \cr
  0  \cr
  .\cr
  .\cr
  .
}
\references{
Edwards, W. R. and Eberhardt, L. (1967). Estimating cottontail abundance from livetrapping data. Journal of Wildlife Management, 31, 87-96.
}
\keyword{datasets}
