\name{ChaoSpeciesDataInci_raw}
\alias{ChaoSpeciesDataInci_raw}
\docType{data}
\title{
  Cotton Tail Rabbit Data for ChaoSpecies
}
\description{
Live trapping was conducted for 18 consecutive nights. Marking or tagging is needed to distinguish individuals. Often, a unique tag was attached to each individual so that the capture history of each individual captured in the experiment was known. A total of 76 distinct species were found in 18 trapping nights. The incidence-raw data consist of a capture/non-capture matrix (where 1 means a capture and 0 means a non-capture) with 76 rows and 18 columns.
}
\usage{data("ChaoSpeciesDataInci_raw")}
\format{
  The data input format (a data.frame with 76 rows and 18 columns): \cr\cr
  1 1 1 1 0 . . . \cr
  1 0 0 0 0 . . . \cr
  1 1 1 0 0 . . . \cr
  1 0 1 0 0 . . . \cr
  1 0 0 0 1 . . . \cr
  . . . . .       \cr
  . . . . .        \cr
  . . . . . 

}
\references{
Edwards, W. R. and Eberhardt, L. (1967). Estimating cottontail abundance from livetrapping data. Journal of Wildlife Management, 31, 87-96.
}
\keyword{datasets}
