\name{SimilarityPairDataAbu}
\alias{SimilarityPairDataAbu}
\docType{data}
\title{
 Rain forest data for SimilarityPair
}
\description{
 The data was discussed in Chao et al. (2005, 2008). It includes the frequencies from three assemblages: seedlings, saplings and  trees. In this data, we choose seedlings and trees for demo. For each species (row), the entry of the first column denotes the species frequency (or abundance) of the sample taken from the seedlings assemblage, and the second column denotes the species frequency (or abundance) of the sample taken from the trees assemblage. 
}
\usage{data(SimilarityPairDataAbu)}
\format{
 The data input format (a matrix with 120 rows and 2 columns) is:\cr\cr
  17  7\cr
  14  0\cr
 121 31\cr
   .  .\cr  
   .  .\cr
   .  .\cr
   1  0 
}
\references{
  Chao, A., Chazdon, R. L., Colwell, R. K. and Shen, T.-J. (2005). A new statistical approach for assessing similarity of species composition with incidence and abundance data. Ecology Letters 8, 148-159.\cr\cr
  Chao, A., Jost, L., Chiang, S. C., Jiang, Y.-H. and Chazdon, R. (2008). A Two -stage probabilistic approach to multiple-community similarity indices. Biometrics 64, 1178-1186.
}

\keyword{datasets}
