\name{SimilarityPairDataInci}
\alias{SimilarityPairDataInci}
\docType{data}
\title{
 Ant Data for SimilarityPair
}
\description{
 The data was collected in Costa-Rica (Longino et al. 2002). It includes incidence frequencies of tropical rainforest ants using three sampling techniques: (a) Berlese extraction of soil samples (217 samples), (b) fogging samples from canopy fogging (459 samples), and (c) Malaise trap samples for flying and crawling insects (62 samples). In this data, we choose (a) and (c) for demo. The first row denotes the number of sampling units 217 and 62, respectively. Beginning with the second row, each row denotes the total number of presences of each species in all the samples. In this demo data, the first species was observed 1 and 0, respectively. The second species was observed 0 and 3, respectively, and so on.
}
\usage{data(SimilarityPairDataInci)}
\format{
 The data input format (a matrix with 277 rows and 2 columns) is:\cr\cr
  217 62  \cr
   1   0  \cr
   0   3  \cr
   .   .  \cr  
   .   .  \cr
   .   .  \cr
   0   0 

}
\references{
 Longino, J. T., Coddington, J. and Colwell, R. K. (2002). The ant fauna of a tropical rain forest: estimating species richness three different ways. Ecology 83, 689-702.
}
\keyword{datasets}
