\name{SimilarityPairDataInci_raw}
\alias{SimilarityPairDataInci_raw}
\docType{data}
\title{
Soil Ciliates Data for SimilarityPair
}
\description{
A total of 51 soil samples were taken from three areas ( EtoshaPan, CentralNamibDesert, SouthernNamibDesert) of Namibia. In area EtoshaPan, there were 19 soil samples and the number of observed species was 216. In area CentralNamibDesert, there were 17 soil samples and the number of observed species was 130. In area SouthernNamibDesert, there were 15 soil samples and the number of observed species was 150. We choose the first two areas data for demo.
}
\usage{data("SimilarityPairDataInci_raw")}
\format{
A list 2: \cr
For the first list (\code{SimilarityPairDataInci_raw[[1]]}) , a matrix with 365 rows and 19 columns; \cr
For the second list (\code{SimilarityPairDataInci_raw[[2]]}), a matrix with 365 rows and 17 columns.\cr\cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  . . . . .       \cr
  . . . . .       \cr
  . . . . .       \cr\cr

  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  0 0 0 0 0 . . . \cr
  . . . . .       \cr
  . . . . .       \cr
  . . . . .       
}
\references{
Foissner, W., Agatha, S., & Berger, H. (2002) Soil Ciliates (Protozoa, Ciliophora) From Namibia (Southwest Africa), With Emphasis on Two Contrasting Environments, the Etosha Region and the Namib Desert. Denisia, 5, 1-1459.
}
\keyword{datasets}
