% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spader.R
\name{ChaoShared}
\alias{ChaoShared}
\title{Estimation of the number of shared species in two assemblages}
\usage{
ChaoShared(data, datatype = c("abundance", "incidence"), se = TRUE,
  nboot = 200, conf = 0.95)
}
\arguments{
\item{data}{a matrix, data.frame (species by sites) of species abundances or incidence frequencies with two columns. 
If \code{datatype = "incidence"}, then the first entry of the input data must be the total number of sampling units, 
followed by species incidence frequencies in each column (See example \code{data(ChaoSharedDataInci)}).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or 
sampling-unit-based incidence data (\code{datatype = "incidence"}).}

\item{se}{a logical variable to calculate the bootstrap standard error and the associated confidence interval.}

\item{nboot}{an integer specifying the number of bootstrap replications.}

\item{conf}{a positive number \eqn{\le} 1 specifying the level of confidence interval.}
}
\value{
a list of two objects: \code{$BASIC_DATA_INFORMATION} for summarizing data information;\cr
\code{$ESTIMATION_RESULTS_OF_THE_NUMBER_OF_SHARED_SPECIES} for showing a table of various shared richess estimates, standard errors, and the associated confidence intervals.
}
\description{
\code{ChaoShared}: estimation of shared species richness in two assemblages based on the following two sampling schemes. \cr
\enumerate{
  \item{(Abundance data) In each assemblage, a random sample of individuals is taken and species abundances/frequencies are recorded.}
  \item{(Incidence data) Each assemblage is sampled several times and species presence/absence data for multiple sampling units are recorded.}
}
}
\examples{
data(ChaoSharedDataAbu)
ChaoShared(ChaoSharedDataAbu, datatype="abundance", se=TRUE, nboot=200, conf=0.95)
data(ChaoSharedDataInci)
ChaoShared(ChaoSharedDataInci, datatype="incidence", se=TRUE, nboot=200, conf=0.95)
}
\references{
Chao, A., Hwang, W.-H., Chen, Y.-C. and Kuo. C.-Y. (2000). Estimating the  number of shared species in two communities. Statistica Sinica 10, 227-246.\cr\cr
Chao, A., Shen, T.-J. and Hwang, W.-H. (2006). Application of Laplace boundary-mode approximations to estimate species and shared species richness.  Australian and New Zealand Journal of Statistics 48, 117-128.\cr\cr
Pan, H. Y., Chao, A. and Foissner, W. (2009). A non-parametric lower bound for the number of species shared by multiple communities. Journal of Agricultural, Biological and Environmental Statistics 14, 452-468.
}

