% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spader.R
\name{ChaoShared}
\alias{ChaoShared}
\title{Estimation of the number of shared species in two communities}
\usage{
ChaoShared(data, datatype = c("abundance", "incidence"), se = TRUE,
  nboot = 200, conf = 0.95)
}
\arguments{
\item{data}{a numerical matrix or a data frame with two columns. Each column represent the species abundances or incidence frequencies of each community.
If \code{datatype = "incidence"}, then the input format of first entry should be total number of sampling units, and followed by species incidence frequencies in each column (See examples).}

\item{datatype}{the data type of input data. That is individual-based abundance data (\code{datatype = "abundance"}) or presence/absence sample-base incidence data (\code{datatype = "incidence"}).}

\item{se}{calculate bootstrap standard error and show confidence interval; default is \code{TRUE}.}

\item{nboot}{the number of bootstrap resampling times, default is \code{200}.}

\item{conf}{a confidence coefficient between 0 and 1.}
}
\value{
returns a list of two. First is the basic data informaiton; the other is a table of various estimators, their standard error and \code{100*conf} \% confidence interval.
}
\description{
\code{ChaoShared} is a function to provide a shared species estimator in tow communities based on the following two sampling schemes. \cr
\enumerate{
  \item{In each community, a random sample of individuals is taken and species frequencies or abundances are recored;}
  \item{Each communtiy is sample several times or the whole area is divided into several quadrats and species presence/absence data for multiple samples/quadrats are recorded.
  }
}
}
\examples{
# load the individual-base (abundance) data
data(SharedSpecDemoAbu)
# Estimation of shared species
ChaoShared(SharedSpecDemoAbu, datatype="abundance", se=TRUE, nboot=200, conf=0.95)
# load the presence/absence sample-base (incidence) data
data(SharedSpecDemoInci)
# Estimation of shared species
ChaoShared(SharedSpecDemoInci, datatype="incidence", se=TRUE, nboot=200, conf=0.95)
}
\author{
Anne Chao, K. H. Ma and T. C. Hsieh
}

