% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spader.R
\name{ChaoSpecies}
\alias{ChaoSpecies}
\title{Etimation of species richness}
\usage{
ChaoSpecies(data, datatype = c("abundance", "incidence"), k = 10,
  conf = 0.95)
}
\arguments{
\item{data}{a vector of species sample frequencies (for abundance data), or incidence-based sample frequencies (1st entry must be the number of sampling unit).}

\item{datatype}{a character of data type,"abundance" or "incidence".}

\item{k}{cut-off point. It is a value that separates frequency counts into abundant and rare groups}

\item{conf}{a confidence coefficient between 0 and 1.}
}
\value{
a vector of species richness estimator and its confidence interval.
}
\description{
\code{ChaoSpecies} is a function to estimate species richness in one community.
}
\examples{
data(SpecDemoAbu)
ChaoSpecies(SpecDemoAbu, datatype="abundance", k = 10, conf = 0.95)
data(SpecDemoInci)
ChaoSpecies(SpecDemoInci, datatype="incidence", k = 10, conf = 0.95)
}
\author{
Anne Chao, K. H. Ma and T. C. Hsieh
}

