\name{DivDemoAbu}
\alias{DivDemoAbu}
\docType{data}
\title{
 Old-growth rain forest data 
}
\description{
 The original data were collected by Dr. Robin Chazdon and colleagues and  
 discussed in Chao et al. (2008).
}
\usage{data(DivDemoAbu)}
\format{
 The format is:\cr\cr
  17 \cr
  14 \cr
 121 \cr
   . \cr  
   . \cr
   . \cr
   1 \cr
   1 \cr
   1
}

\references{
  Chao, A., Ma, K. H. and Hsieh, T. C. (2015) Online Program SpadeR   
  (Species Prediction And Diversity Estimation in R). Program and Users 
  Guide published at http://chao.stat.nthu.edu.tw/spader_guide.pdf.
}

\keyword{datasets}
