% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spader.R
\name{Genetic}
\alias{Genetic}
\title{Estimation of Genetic measure}
\usage{
Genetic(X, q = 2, nboot = 200)
}
\arguments{
\item{X}{a matrix (species by sites) of species abundances.}

\item{q}{set diversity order \code{q = 0}, \code{Genetic} computes the estimates of Sorensen index for pairwise assemblages; 
set diversity order \code{q = 1}, computes the estimates of Horn index for pairwise assemblages; 
set diversity order \code{q = 2}, computes the estimates of Morisita index for pairwise assemblages. 
For diversity order q = 0, 1, 2, \code{Genetic}  computes the overlap estimates among all assemblages.}

\item{nboot}{an integer specifying the number of bootstrap replications.}
}
\value{
a list of five objects: \code{$info} for summarizing data information;
\code{$overlap} for showing the C0N, C1N and C2N of this data;
\code{$pairwise} for showing the pairwise data CqN output;
\code{$similarity.matrix} for showing pairwise similarity of this data. \code{q} you choose.
}
\description{
\code{Genetic}: estimation of the multiple community similarity C_qN and dissimilarity 1-C_qN including Jost (2008) differentiation measure D.
}
\examples{
data(GeneticDataAbu)
Genetic(GeneticDataAbu, q=2, nboot=200)
}
\references{
Chao, A., Jost, L., Chiang, S. C., Jiang, Y.-H. and Chazdon, R. (2008). A Two-stage probabilistic approach to multiple-community similarity indices. Biometrics, 64, 1178-1186.\cr\cr
Jost, L. (2008). GST and its relatives do not measure differentiation. Molecular Ecology, 17, 4015-4026.
}

