\name{MulComSimDemoAbu}
\alias{MulComSimDemoAbu}
\docType{data}
\title{
 Rain forest data
}
\description{
 The data include the frequencies from three assemblages: seedlings (column 1), saplings (column 2) and 
 trees (column 3) in the LEP old-growth rain forest.
}
\usage{data(MulComSimDemoAbu)}
\format{
 The format is:\cr\cr
  17 48  7\cr
  14 38  0\cr
 121 37 31\cr
   .  .  .\cr  
   .  .  .\cr
   .  .  .\cr
   1  0  0
}

\references{
  Chao, A., Ma, K. H. and Hsieh, T. C. (2015) Online Program SpadeR   
  (Species Prediction And Diversity Estimation in R). Program and 
  Users Guide published at http://chao.stat.nthu.edu.tw/spader_guide
  .pdf.
}

\keyword{datasets}
