\name{SharedSpecDemoAbu}
\alias{SharedSpecDemoAbu}
\docType{data}
\title{
 Birds data
}
\description{
  The data is bird frequency records of two estuaries.
  The first column denotes the frequency (or abundance) of a species 
  discovered in estuary I, and the second column denotes the 
  frequency (or abundance) of the same species in estuary II.
}
\usage{data(SharedSpecDemoAbu)}
\format{
 The format is:\cr\cr
 39 42\cr
  0 70\cr
  0 1\cr
  . .\cr  
  . .\cr
  . .\cr
  1 1 
}

\references{
  Chao, A., Ma, K. H. and Hsieh, T. C. (2015) Online Program SpadeR   
  (Species Prediction And Diversity Estimation in R). Program and  
  Users Guide published at http://chao.stat.nthu.edu.tw/spader_guide
  .pdf.
}

\keyword{datasets}
