\name{ChaoSharedDataInci}
\alias{ChaoSharedDataInci}
\docType{data}
\title{
 The Hong Kong Big Bird Race Data for ChaoShared
}
\description{
 The data is bird incidence frequencies of two assemblages. The first row denotes the number of sampling units 16 and 17, respectively. Beginning with the second row, each row denotes the total number of detections of a given species in all the sampling units. The first species entered as (1, 0) was observed by only one team in 2015 and was not observed by any team in 2016. The second species entered as (3, 8) was observed by three teams in 2015 and was observed by eight teams in 2016. Finally, the species in the last row was not observed by any team in 2015 and was observed by only one team in 2016.
 }
\usage{data(ChaoSharedDataInci)}
\format{
 The data input format (a data.frame with 281 rows and 2 columns) is:\cr\cr
 16 17 \cr
  1  0 \cr
  3  8 \cr
  .  . \cr
  .  . \cr
  .  . \cr
  0  1
}
\references{
WWF-Hong Kong's longest-running fundraising event. Established in 1984.
}
\keyword{datasets}
