\name{SharedSpecDemoInci}
\alias{SharedSpecDemoInci}
\docType{data}
\title{
 Birds data
}
\description{
 The Hong Kong Big Bird Race (BBR) is an annual competition among teams of  
 birdwatchers. The challenge is to record as many bird species as possible during 
 a fixed interval of time in the Hong Kong territory. The data is 19 teams 
 competed in 1999 while 20 teams competed in 2000.
}
\usage{data(SharedSpecDemoInci)}
\format{
 The format is:\cr\cr
 19 20\cr
  0 0 \cr
  0 0 \cr
  . . \cr  
  . . \cr
  . . \cr
 19 20 
}

\references{
  Chao, A., Ma, K. H. and Hsieh, T. C. (2015) Online Program SpadeR   
  (Species Prediction And Diversity Estimation in R). Program and Users 
  Guide published at http://chao.stat.nthu.edu.tw/spader_guide.pdf.
}

\keyword{datasets}
