% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spader.R
\name{SimilarityMult}
\alias{SimilarityMult}
\title{Estimation of multiple-assemblage similarity measure}
\usage{
SimilarityMult(X, q = 2, nboot = 200)
}
\arguments{
\item{X}{a matrix (species by sites) of species abundances.}

\item{q}{set diversity order \code{q = 0}, \code{SimilarityMult} computes the estimates of Sorensen index for pairwise assemblages;
set diversity order \code{q = 1}, computes the estimates of Horn index for pairwise assemblages;
set diversity order \code{q = 2}, computes the estimates of Morisita index for pairwise assemblages.
For diversity order q = 0, 1, 2, \code{SimilarityMult}  computes the overlap estimates among all assemblages.}

\item{nboot}{an integer specifying the number of bootstrap replications.}
}
\value{
a list of six objects: \code{$info} for summarizing data information; \code{$overlap} for showing overlap estimates along with related statistics among all assemblages;
\code{$pairwise} and {$similarity.matrix} for showing pairwise similarity estimates (with related statistics) depended on which diversity order \code{q} you choose;
\code{$method}: fixed \code{method="absolute"}; \code{$q} for showing which diversity order \code{q} you choose.
}
\description{
\code{SimilarityMult}: estimation of the generalized Sorensen, Horn, and Morisita similarity/dissimilarity indices for comparing frequency or abundance data taken from more than two assemblages.
}
\examples{
data(SimilarityMultDataAbu)
SimilarityMult(SimilarityMultDataAbu,q=2,nboot=200)
}
\references{
Chao, A., Jost, L., Chiang, S. C., Jiang, Y.-H. and Chazdon, R. (2008). A Two- stage probabilistic approach to multiple-community similarity indices. Biometrics, 64, 1178-1186.\cr\cr
Jost, L. (2008). GST and its relatives do not measure differentiation. Molecular Ecology, 17, 4015-4026.
}

