% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spader.R
\name{SimilarityMult}
\alias{SimilarityMult}
\title{Etimation of multiple-community similarity measure}
\usage{
SimilarityMult(X, q = 2, nboot = 200)
}
\arguments{
\item{X}{a matrix of species sample frequencies (for abundance data), row is species number, and column is community number.}

\item{q}{For q=0, the SpadeR computes the estimates of Sorensen index for paired-wise communities. For q=1, the SpadeR computes the estimates of 1-Horn index for paired-wise communities. For q=2, the SpadeR computes the estimates of Morisita index for paired-wise communities.}

\item{nboot}{the number of bootstrap resampling times, default is 200.}
}
\value{
include every community data summary, overlap measure in all communities, and pair-wise similarity measure depended on which order you choose in two communities.
}
\description{
\code{SimilarityMult} This part computes the generalized Sorensen, Horn similarity, and Morisita similarity/dissimilarity indices for comparing frequency or abundance data from more than two communities.
}
\examples{
data(MulComSimDemoAbu)
SimilarityMult(MulComSimDemoAbu,q=2,nboot=200)
}
\author{
Anne Chao, K. H. Ma and T. C. Hsieh
}

