% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spader.R
\name{SimilarityMult}
\alias{SimilarityMult}
\title{Estimation of multiple-community similarity measures}
\usage{
SimilarityMult(
  data,
  datatype = c("abundance", "incidence_freq", "incidence_raw"),
  units,
  q = 2,
  nboot = 200,
  goal = "relative",
  weight_type = c("Size Weight", "Equal Weight", "Others"),
  weight
)
}
\arguments{
\item{data}{a data.frame of species abundances/incidences.\cr}

\item{datatype}{type of input data, "abundance", "incidence_freq" or "incidence_raw". \cr}

\item{units}{number of sampling units in each community. For \code{datatype = "incidence_raw"}, users must specify the number of sampling units taken from each community. This argument is not needed for "abundance" and "incidence_freq" data. \cr}

\item{q}{a specified order to use to compute pairwise similarity measures.}

\item{nboot}{an integer specifying the number of bootstrap replications.}

\item{goal}{a specified estimating goal to use to compute pairwise similarity measures:comparing species relative abundances (\code{goal=relative}) or comparing species absolute abundances (\code{goal=absolute}). \cr\cr}

\item{weight_type}{choose type of weight for comparing relative abundance, \code{Size Weight} for using sample size of each community as weight, \code{Equal Weight} 
for using same weight for each community, \code{Others} for using the weight user specified.}

\item{weight}{only needed when \code{weight_type=Others}, users must specify the weight for each community.}
}
\value{
a list of ten objects: \cr\cr
\code{$datatype} for showing the specified data types (abundance or incidence).\cr\cr
\code{$info} for summarizing data information.\cr\cr 
\code{$Empirical_absolute} for showing the observed values of the absolute abundance-based similarity indices 
include the richness-based Jaccard and Sorensen, Horn, Morisita-Horn, regional-overlap indices. \cr\cr
\code{$Empirical_relative} for showing the observed values of the relative abundance-based similarity indices
include the size-weighted Sorenson and Jaccard, size-weighted Horn, size-weighted Morisita-Horn, size-weighted regional-overlap. \cr \cr
\code{$Estimated_absolute} for showing the estimaed values of the absolute abundance-based similarity indices 
include the richness-based Jaccard and Sorensen, Horn, Morisita-Horn, regional-overlap indices. \cr\cr
\code{$Estimated_relative} for showing the estimated values of the relative abundance-based similarity indices
include the size-weighted Sorenson and Jaccard, size-weighted Horn, size-weighted Morisita-Horn, size-weighted regional-overlap. \cr\cr
\code{$pairwise} and \code{$similarity.matrix} for showing respectively the pairwise dis-similarity
estimates (with related statistics) and the similarity matrix for various measures depending on the
diversity order \code{q} and the \code{goal} specified in the function arguments. \cr\cr
\code{$goal} for showing the goal specified in the argument goal (absolute or relative) used to compute pairwise similarity.\cr\cr
\code{$q} for showing which order \code{q} specified to compute pairwise similarity. \cr\cr
}
\description{
\code{SimilarityMult}: Estimation various \eqn{N}-community similarity indices. The absolute abundance-based
indices include the richness-based Jaccard and Sorensen, Horn, Morisita-Horn, regional-overlap indices. The relative-based indices include 
the size-weighted Sorenson and Jaccard, size-weighted Horn, size-weighted Morisita-Horn, size-weighted regional-overlap. 
Three types of data are supported: Type (1)
abundance data (datatype="abundance"), Type (2) incidence-frequency data
(datatype="incidence_freq"), and Type (2B) incidence-raw data (datatype="incidence_raw"); see
\code{SpadeR-package} details for data input formats.
}
\examples{
\dontrun{
data(SimilarityMultData)
# Type (1) abundance data 
SimilarityMult(SimilarityMultData$Abu,"abundance",q=2,nboot=200,"relative")
# Type (2) incidence-frequency data 
SimilarityMult(SimilarityMultData$Inci,"incidence_freq",q=2,nboot=200,"relative")
# Type (2B) incidence-raw data 
SimilarityMult(SimilarityMultData$Inci_raw,"incidence_raw",
units=c(19,17,15),q=2,nboot=200,"relative")
}
}
\references{
Chao, A., and Chiu, C. H. (2016). Bridging the variance and diversity decomposition approaches to beta diversity via similarity and differentiation measures. Methods in Ecology and Evolution, 7, 919-928. \cr\cr
Chao, A., Jost, L., Hsieh, T. C., Ma, K. H., Sherwin, W. B. and Rollins, L. A. (2015). Expected Shannon entropy and Shannon differentiation between subpopulations for neutral genes under the finite island model. Plos One, 10:e0125471.\cr\cr
Chiu, C. H., Jost, L. and Chao, A. (2014). Phylogenetic beta diversity, similarity, and differentiation measures based on Hill numbers. Ecological Monographs, 84, 21-44.\cr\cr
Gotelli, N. G. and Chao, A. (2013). Measuring and estimating species richness, species diver- sity,
and biotic similarity from sampling data. Encyclopedia of Biodiversity, 2nd Edition, Vol. 5, 195-211, Waltham, MA.
}
