% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spader.R
\name{SimilarityPair}
\alias{SimilarityPair}
\title{Estimation of two-assemblage similarity index}
\usage{
SimilarityPair(X, datatype = c("abundance", "incidence"), nboot = 200)
}
\arguments{
\item{X}{a matrix (species by sites) of species abundances, or incidence frequencies with two columns.
If \code{datatype = "incidence"}, then the first entry of the input data must be total number of sampling units in each column.
(See example \code{data(SimilarityPairDataInci)}).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or
sampling-unit-based incidence data (\code{datatype = "incidence"}).}

\item{nboot}{an integer specifying the number of replications.}
}
\value{
a list of six objects: \code{$datatype} for showing data type;
\code{$info1} and \code{$info2} for summarizing data information;
\code{$similarity} for showing the various similarity index estimates along with related statistics;
\code{$assemblage1} and {$assemblage2} for showing species richness estimates with related statistics of each assemblage.
}
\description{
\code{SimilarityPair}: estimation of various similarity indices based on either individual-based abundance data or sampling-unit-based incidence data
taken from two assemblages. The incidence-based indices include the classic Jaccard, Sorensen and Lennon et al. (2001) indices,
and the abundance-based indices include the Bray-Curtis, Morisita-Horn, Horn and abundance-based Jaccard
and Sorensen indices developed in Chao et al. (2005).
}
\examples{
data(SimilarityPairDataAbu)
SimilarityPair(SimilarityPairDataAbu, datatype="abundance")
data(SimilarityPairDataInci)
SimilarityPair(SimilarityPairDataInci,datatype="incidence")
}
\references{
Chao, A. (1984). Nonparametric estimation of the number of classes in a population. Scandinavian Journal of Statistics 11, 265-270.\cr\cr
Chao, A. (1987). Estimating the population size for capture-recapture data with unequal catchability. Biometrics 43, 783-791.\cr\cr
Chao, A. (2005). Species estimation and applications. Encyclopedia of Statistical Sciences, Second Edition, Vol. 12, 7907-7916 (N. Balakrishnan, C. B. Read and B. Vidakovic, Editors), Wiley, New York.\cr\cr
Chao, A., Chazdon, R. L., Colwell, R. K. and Shen, T.-J. (2005). A new statistical approach for assessing similarity of species composition with incidence and abundance data. Ecology Letters 8, 148-159.\cr\cr
Chao, A., Chazdon, R. L., Colwell, R. K. and Shen, T.-J. (2006). Abundance-based similarity indices and their estimation when there are unseen species in samples. Biometrics, 62, 361-371.\cr\cr
Chao, A., Hwang, W.-H., Chen, Y.-C. and Kuo. C.-Y. (2000). Estimating the number of shared species in two communities. Statistica Sinica 10, 227-246.\cr\cr
Chao, A. and Lee, S.-M. (1992). Estimating the number of classes via sample coverage. Journal of the American Statistical Association 87, 210-217.\cr\cr
Chao, A., Ma, M-C. and Yang, M. C. K. (1993). Stopping rule and estimation for recapture debugging with unequal detection rates. Biometrika 80, 193-201.\cr\cr
Chao, A., Shen, T.-J. and Hwang, W.-H. (2006). Application of Laplace boundary-mode approximations to estimate species and shared species richness. Australian and New Zealand Journal of Statistics 48, 117-128.\cr\cr
Lee, S.-M. and Chao, A. (1994). Estimating population size via sample coverage for closed capture-recapture models. Biometrics 50, 88-97.\cr\cr
Lennon, J. J., Koleff, P., Greenwood, J. J. D. and Gaston, K. J. (2001). The geographical structure of British bird distributions: diversity, spatial turnover and scale. J. Anim. Ecol., 70, 966-979.\cr\cr
Shen, T.-J. (2003). Prediction of Biodiversity, Ph.D. dissertation, National Tsing-Hua University, HsinChu, Taiwan.\cr\cr
Shen, T.-J., Chao, A. and Lin, J.-F. (2003). Predicting the number of new species in further taxonomic sampling. Ecology 84, 798-804.
}

