% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spader.R
\name{SimilarityPair}
\alias{SimilarityPair}
\title{Etimation of two-community similarity index}
\usage{
SimilarityPair(X, datatype = c("abundance", "incidence"), nboot = 200)
}
\arguments{
\item{X}{a matrix of species sample frequencies (for abundance data), or incidence-based sample frequencies (1st row must be the number of sampling unit) for two communities.}

\item{datatype}{a character of data type,"abundance" or "incidence".}

\item{nboot}{the number of bootstrap resampling times, default is 200.}
}
\value{
include two communities data summary and the various similarity indice estimation
}
\description{
\code{SimilarityPair} This part features various similarity indices for comparing data from two assemblages based on either multiple-sample incidence data or abundance data. The incidence-based indices include the classic Jaccard, Sorensen and Lennon et al. (2001) indices, and the abundance-based indices include the Bray-Curtis, Morisita-Horn, Horn and two newly developed abundance-based Jaccard and Sorensen indices.
}
\examples{
data(TwoComSimDemoAbu)
SimilarityPair(TwoComSimDemoAbu, datatype="abundance")
data(TwoComSimDemoInci)
SimilarityPair(TwoComSimDemoInci,datatype="incidence")
}
\author{
Anne Chao, K. H. Ma and T. C. Hsieh
}

