\name{SpadeR-package}
\alias{SpadeR-package}
\alias{SpadeR}
\docType{package}
\title{Species prediction and diversity estimation with R

}
\description{
This package computes various biodiversity indices based on individual-based (abundance) data or sampling-unit-based (incidence) data taken from one or multiple assemblages. It contains six parts: 1. ChaoShared (estimating the number of shared species for two assemblages); 2. ChaoSpecies (estimating species richness for one assemblage); 3. Diversity (estimating a continuous diversity profile including species richness, Shannon index/diversity and Simpson index/diversity); 4. SimilarityMult (estimating Sorensen, Horn, and Morisita similarity indices for more than two assemblages); 5. SimilartyPair (estimating various similarity indices for two assemblages). The sampling-unit-based indices include the class Jaccard, Sorensen and Lennon et al. indices; the individual-based indices include the Bray-curtis, Morisita-Horn, Horn and abundance-based Jaccard and Sorensen indices; 6. Genetic (estimating allelic differentiation/similarity among subpopulations). This part is almost identical to SimilarityMult except that it is focused on genetic applications.
}
\details{
\tabular{ll}{
Package: \tab SpadeR\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2016-05-09\cr
License: \tab GPL (>= 3)\cr
}
functions:  ChaoShared, ChaoSpecies, Diversity, SimilarityMult, SimilarityPair, Genetic
}
\author{
Anne Chao, K. H. Ma and T. C. Hsieh 

Maintainer: K. H. Ma <smallmar123@gmail.com>

}
\keyword{ package }


