\name{SpadeR-package}
\alias{SpadeR-package}
\alias{SpadeR}
\docType{package}
\title{Species prediction and diversity estimation with R

}
\description{
SpadeR computes various biodiversity indices based on different types of sample data from one or multiple communities.
}
\details{
\tabular{ll}{
Package: \tab SpadeR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-07\cr
License: \tab GPL (>= 3)\cr
}
}
\author{
Anne Chao, K. H. Ma and T. C. Hsieh 

Maintainer: K. H. Ma <smallmar123@gmail.com>

}
\keyword{ package }


