\name{SpecDemoAbu}
\alias{SpecDemoAbu}
\docType{data}
\title{
 Bird species abundance frequency counts
}
\description{
 A set of birds data from Magurran (1988)
}
\usage{data(SpecDemoAbu)}
\format{
  The format is:\cr\cr
  752 \cr
  276 \cr
  194 \cr
   .  \cr
   .  \cr
   .  \cr
   1  \cr
   1  \cr
   1
}

\references{
  Chao, A., Ma, K. H. and Hsieh, T. C. (2015) Online Program SpadeR   
  (Species Prediction And Diversity Estimation in R). Program and Users 
  Guide published at http://chao.stat.nthu.edu.tw/spader_guide.pdf.
}

\keyword{datasets}
