\name{SpecDemoInci}
\alias{SpecDemoInci}
\docType{data}
\title{
  Soil samples
}
\description{
  A set of soil samples of seedlings data
}
\usage{data(SpecDemoInci)}
\format{
 The format is:\cr\cr
  121 \cr
   1  \cr
   1  \cr
   1  \cr  
   .  \cr
   .  \cr
   .  \cr
  47  \cr
  52  \cr
  61
}

\references{
  Chao, A., Ma, K. H. and Hsieh, T. C. (2015) Online Program SpadeR   
  (Species Prediction And Diversity Estimation in R). Program and 
  Users Guide published at http://chao.stat.nthu.edu.tw/spader_guide
  .pdf.
}

\keyword{datasets}
