\name{TwoComSimDemoAbu}
\alias{TwoComSimDemoAbu}
\docType{data}
\title{
 Rain forest data
}
\description{
 The data include the frequencies from two assemblages: seedlings 
 (column 1), and trees (column 2) sampled in an old-growth rain  
 forest named LEP.
}
\usage{data(TwoComSimDemoAbu)}
\format{
 The format is:\cr\cr
  17  7\cr
  14  0\cr
 121 31\cr
   .  .\cr  
   .  .\cr
   .  .\cr
   1  0 
}

\references{
  Chao, A., Ma, K. H. and Hsieh, T. C. (2015) Online Program SpadeR   
  (Species Prediction And Diversity Estimation in R). Program and 
  Users Guide published at http://chao.stat.nthu.edu.tw/spader_guide
  .pdf.
}

\keyword{datasets}
