\name{TwoComSimDemoInci}
\alias{TwoComSimDemoInci}
\docType{data}
\title{
 Ant data
}
\description{
 The data include incidence frequencies of tropical rainforest ants 
 using two sampling techniques: (a) Berlese extraction of soil 
 samples (217 samples) and (b) Malaise trap samples for flying and 
 crawling insects (62 samples)
}
\usage{data(TwoComSimDemoInci)}
\format{
 The format is:\cr\cr
 217 62\cr
   1  0\cr
   0  3\cr
   .  .\cr  
   .  .\cr
   .  .\cr
   0  0 
}

\references{
   Chao, A., Ma, K. H. and Hsieh, T. C. (2015) Online Program SpadeR   
  (Species Prediction And Diversity Estimation in R). Program and 
  Users Guide published at http://chao.stat.nthu.edu.tw/spader_guide
  .pdf.
}

\keyword{datasets}
